requirements.txt
numpy
opencv-python
Pillow
torch>=1.7
torchvision>=0.8.0
tqdm
huggingface-hub
pyTelegramBotAPI
python-dotenv